refactor: move keep awake to DeviceService
