+++ Runtime/Device/DeviceService.cs
@@ -1,5 +1,6 @@
 using System;
 using GameLovers.MobileServices.Device.Internal;
+using UnityEngine;
 
 // ReSharper disable once CheckNamespace
 namespace GameLovers.MobileServices.Device
@@ -7,10 +8,19 @@ namespace GameLovers.MobileServices.Device
 	/// <inheritdoc />
 	public sealed class DeviceService : IDeviceService, IDisposable
 	{
+		/// <summary>
+		/// Controls whether the device screen stays awake. When <c>true</c>, sets
+		/// <c>Screen.sleepTimeout</c> to <c>SleepTimeout.NeverSleep</c>; when <c>false</c>,
+		/// restores <c>SleepTimeout.SystemSetting</c>.
+		/// </summary>
+		public static bool KeepAwake
+		{
+			get => Screen.sleepTimeout == SleepTimeout.NeverSleep;
+			set => Screen.sleepTimeout = value ? SleepTimeout.NeverSleep : SleepTimeout.SystemSetting;
+		}
+
 		/// <inheritdoc />
 		public ISafeAreaService SafeArea { get; }
-		/// <inheritdoc />
-		public IScreenWakeService ScreenWake { get; }
 		/// <inheritdoc />
 		public IBatteryService Battery { get; }
 		/// <inheritdoc />
@@ -26,7 +36,6 @@ namespace GameLovers.MobileServices.Device
 
 		public DeviceService(
 			ISafeAreaService safeArea,
-			IScreenWakeService screenWake,
 			IBatteryService battery,
 			IIosAudioSessionService audioSession,
 			IPermissionsService permissions,
@@ -34,7 +43,6 @@ namespace GameLovers.MobileServices.Device
 			IDeepLinkService deepLink)
 		{
 			SafeArea = safeArea;
-			ScreenWake = screenWake;
 			Battery = battery;
 			AudioSession = audioSession;
 			Permissions = permissions;
@@ -45,20 +53,19 @@ namespace GameLovers.MobileServices.Device
 		// Tuple-routed delegating ctor so the host-dependent children share one explicit host
 		// instance constructed up-front, not separate accesses to the singleton during a
 		// constructor chain (cleaner ownership signal in the umbrella's call stack).
-		private DeviceService((ISafeAreaService, IScreenWakeService, IBatteryService,
+		private DeviceService((ISafeAreaService, IBatteryService,
 			IIosAudioSessionService, IPermissionsService, IAttService, IDeepLinkService) defaults)
 			: this(defaults.Item1, defaults.Item2, defaults.Item3,
-				   defaults.Item4, defaults.Item5, defaults.Item6, defaults.Item7)
+				   defaults.Item4, defaults.Item5, defaults.Item6)
 		{
 		}
 
-		private static (ISafeAreaService, IScreenWakeService, IBatteryService,
+		private static (ISafeAreaService, IBatteryService,
 			IIosAudioSessionService, IPermissionsService, IAttService, IDeepLinkService) BuildDefaults()
 		{
 			var host = DeviceServicesHost.Instance;
 			return (
 				new SafeAreaService(host),
-				new ScreenWakeService(),
 				new BatteryService(host),
 				new IosAudioSessionService(),
 				new PermissionsService(),
