--- Runtime/Device/State/IScreenWakeService.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: bcda1d4b4808e4808ba9b6fd7e980d99
