--- Runtime/Device/State/ScreenWakeService.cs
@@ -1,16 +1,0 @@
-using UnityEngine;
-
-// ReSharper disable once CheckNamespace
-namespace GameLovers.MobileServices.Device
-{
-	/// <inheritdoc />
-	public sealed class ScreenWakeService : IScreenWakeService
-	{
-		/// <inheritdoc />
-		public bool KeepAwake
-		{
-			get => Screen.sleepTimeout == SleepTimeout.NeverSleep;
-			set => Screen.sleepTimeout = value ? SleepTimeout.NeverSleep : SleepTimeout.SystemSetting;
-		}
-	}
-}
