--- Runtime/Device/State/ScreenWakeService.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: f4ef98007e9d54bd2acaddde0c165c80
