+++ Tests/EditMode/Unit/DeviceServiceTest.cs
@@ -2,6 +2,7 @@ using System;
 using GameLovers.MobileServices.Device;
 using NSubstitute;
 using NUnit.Framework;
+using UnityEngine;
 
 // ReSharper disable once CheckNamespace
 
@@ -19,18 +20,18 @@ namespace GameLoversEditor.MobileServices.Tests
 		public interface IDeepLinkServiceDisposable : IDeepLinkService, IDisposable { }
 
 		private ISafeAreaServiceDisposable _safeArea;
-		private IScreenWakeService _screenWake;
 		private IBatteryServiceDisposable _battery;
 		private IIosAudioSessionService _audioSession;
 		private IPermissionsService _permissions;
 		private IAttService _att;
 		private IDeepLinkServiceDisposable _deepLink;
+		private int _originalSleepTimeout;
 
 		[SetUp]
 		public void Init()
 		{
+			_originalSleepTimeout = Screen.sleepTimeout;
 			_safeArea = Substitute.For<ISafeAreaServiceDisposable>();
-			_screenWake = Substitute.For<IScreenWakeService>();
 			_battery = Substitute.For<IBatteryServiceDisposable>();
 			_audioSession = Substitute.For<IIosAudioSessionService>();
 			_permissions = Substitute.For<IPermissionsService>();
@@ -38,14 +39,19 @@ namespace GameLoversEditor.MobileServices.Tests
 			_deepLink = Substitute.For<IDeepLinkServiceDisposable>();
 		}
 
+		[TearDown]
+		public void Cleanup()
+		{
+			Screen.sleepTimeout = _originalSleepTimeout;
+		}
+
 		[Test]
 		// ADMIT: DeviceService's injection ctor could store something other than the supplied instance on a child property.
-		// RCR: DeviceService.cs DeviceService(7-arg) — `Att = att` → `Att = new AttService()` → RED (AreSame fails on Att).
+		// RCR: DeviceService.cs DeviceService(6-arg) — `Att = att` → `Att = new AttService()` → RED (AreSame fails on Att).
 		public void InjectionCtor_StoresEachChildOnMatchingProperty()
 		{
 			var service = new DeviceService(
 				_safeArea,
-				_screenWake,
 				_battery,
 				_audioSession,
 				_permissions,
@@ -53,7 +59,6 @@ namespace GameLoversEditor.MobileServices.Tests
 				_deepLink);
 
 			Assert.AreSame(_safeArea, service.SafeArea);
-			Assert.AreSame(_screenWake, service.ScreenWake);
 			Assert.AreSame(_battery, service.Battery);
 			Assert.AreSame(_audioSession, service.AudioSession);
 			Assert.AreSame(_permissions, service.Permissions);
@@ -68,7 +73,6 @@ namespace GameLoversEditor.MobileServices.Tests
 		{
 			var service = new DeviceService(
 				_safeArea,
-				_screenWake,
 				_battery,
 				_audioSession,
 				_permissions,
@@ -93,7 +97,6 @@ namespace GameLoversEditor.MobileServices.Tests
 
 			var service = new DeviceService(
 				nonDisposableSafeArea,
-				_screenWake,
 				nonDisposableBattery,
 				_audioSession,
 				_permissions,
@@ -101,6 +104,39 @@ namespace GameLoversEditor.MobileServices.Tests
 				nonDisposableDeepLink);
 
 			Assert.DoesNotThrow(service.Dispose);
+		}
+
+		[Test]
+		// ADMIT: DeviceService.KeepAwake = true could fail to prevent the screen from dimming.
+		// RCR: DeviceService.cs KeepAwake.set — true branch `NeverSleep` → `SystemSetting` → RED (expected NeverSleep). 2026-08-13
+		public void KeepAwake_True_SetsScreenSleepTimeoutNeverSleep()
+		{
+			DeviceService.KeepAwake = true;
+
+			Assert.AreEqual(SleepTimeout.NeverSleep, Screen.sleepTimeout);
+		}
+
+		[Test]
+		// ADMIT: DeviceService.KeepAwake = false could leave the screen awake after the application releases the override.
+		// RCR: DeviceService.cs KeepAwake.set — false branch `SystemSetting` → `NeverSleep` → RED (expected SystemSetting). 2026-08-13
+		public void KeepAwake_False_RestoresSystemSetting()
+		{
+			DeviceService.KeepAwake = true;
+			DeviceService.KeepAwake = false;
+
+			Assert.AreEqual(SleepTimeout.SystemSetting, Screen.sleepTimeout);
+		}
+
+		[Test]
+		// ADMIT: DeviceService.KeepAwake could report the inverse of the active screen timeout.
+		// RCR: DeviceService.cs KeepAwake.get — `==` → `!=` → RED (expected true for NeverSleep). 2026-08-13
+		public void KeepAwake_Get_ReflectsScreenSleepTimeout()
+		{
+			Screen.sleepTimeout = SleepTimeout.NeverSleep;
+			Assert.IsTrue(DeviceService.KeepAwake);
+
+			Screen.sleepTimeout = SleepTimeout.SystemSetting;
+			Assert.IsFalse(DeviceService.KeepAwake);
 		}
 	}
 }
