--- Tests/EditMode/Unit/ScreenWakeServiceTest.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: e7c15149e36b14a5fa31081cdf5d45e3
